requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.1.1
pygame==1.9.2a0
scipy==0.17.0
SimpleCV==1.3
Werkzeug==0.11.4
